Add vector collection put and optimize benchmarks
